==> analysis/2.participants.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Created by [Changkun Ou](https://changkun.de).

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

import pandas as pd

df = pd.read_csv('../data/meta.csv')
df = df[df['user_id'] != 12] # exclude Andreas

print(f"male: {len(df[df['gender'] == 0])}, female: {len(df[df['gender'] == 1])}")
print(f"range: [{df['age'].min()}, {df['age'].max()}]")
print(f"mean: {df['age'].mean()}")
print(f"std: {df['age'].std()}")
