==> analysis/3.missing_data.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Based on discussions between [Changkun Ou](https://changkun.de), [Francesco Chiossi](https://www.um.informatik.uni-muenchen.de/personen/mitarbeiter/chiossi/index.html), and [Sven Mayer](https://sven-mayer.com).

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

This notebook aims to identify users with missing data.

import pandas as pd
import numpy as np

df_ecg = pd.read_csv('./clean/ecg.csv')
df_eda = pd.read_csv('./clean/eda.csv')
df_mot = pd.read_csv('./clean/motion.csv')
df_ipq = pd.read_csv('./clean/ipq.csv')
df_tlx = pd.read_csv('./clean/nasa_tlx.csv')
df_pac = pd.read_csv('./clean/paces.csv')

# We should have a total of 3*2*4 = 24 (#methods x #tide x #lap) measures per user.
u1 = df_ecg.groupby(['user_id']).count().query('lap != 24').index.values.tolist()
u2 = df_eda.groupby(['user_id']).count().query('lap != 24').index.values.tolist()
u3 = df_mot.groupby(['user_id']).count().query('lap != 24').index.values.tolist()
# IPQ questionaire have 14 questions and each user is measured 2 times, in total we should have 28 rows per user.
u4 = df_ipq.groupby(['user_id']).count().query('ipq != 28').index.values.tolist()
# NASA-TLX questionaire have 6 questions and each user is measured 6 times, in total we should have 36 rows per user.
u5 = df_tlx.groupby(['user_id']).count().query('tlx != 36').index.values.tolist()
# PACES questionaire have 18 questions and each user is measured 6 times, in total we should have 108 rows per user.
u6 = df_pac.groupby('user_id').count().query('tlx != 108').index.values.tolist()

user_with_missing_data = list(set(u1+u2+u3+u4+u5+u6))
user_with_missing_data

# users that produces NaN in HRV measures
list(set(df_ecg.query('hrv_rmssd != hrv_rmssd')['user_id'].tolist()))
