==> analysis/4.analysis_ecg.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Based on discussions between [Changkun Ou](https://changkun.de), [Francesco Chiossi](https://www.um.informatik.uni-muenchen.de/personen/mitarbeiter/chiossi/index.html), and [Sven Mayer](https://sven-mayer.com).

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

source('stats.R')

df_ecg <- read.csv("./clean/ecg.csv")
# remove users with missing data, see 1.missing.ipynb
df_ecg <- df_ecg %>% filter(!(user_id %in% c(12, 14, 20, 22)))
# replace missing data with median of the hrv_rmssd
df_ecg$hrv_rmssd[is.na(df_ecg$hrv_rmssd)] <- median(df_ecg$hrv_rmssd, na.rm = T)

df_ecg$user_id <- as.factor(df_ecg$user_id)
df_ecg$method <- as.factor(df_ecg$method)
df_ecg$tide <- as.factor(df_ecg$tide)

twoway_anova_test(df_ecg, "ecg_rate")
twoway_anova_test(df_ecg, "hrv_rmssd")

df_ecg <- read.csv("./clean/ecg.csv")
# remove users with missing data, see 1.missing.ipynb
df_ecg <- df_ecg %>% filter(!(user_id %in% c(11, 12, 14, 19, 20, 22, 25)))

df_ecg$user_id <- as.factor(df_ecg$user_id)
df_ecg$method <- as.factor(df_ecg$method)
df_ecg$tide <- as.factor(df_ecg$tide)
twoway_anova_test(df_ecg, "hrv_rmssd")

df_ecg_noins <- df_ecg[df_ecg$method == "NoInstructions",]
df_ecg_gamif <- df_ecg[df_ecg$method == "Gamification",]
df_ecg_imita <- df_ecg[df_ecg$method == "Imitation",]

mean(df_ecg_noins$ecg_rate)
sd(df_ecg_noins$ecg_rate)
mean(df_ecg_gamif$ecg_rate)
sd(df_ecg_gamif$ecg_rate)
mean(df_ecg_imita$ecg_rate)
sd(df_ecg_imita$ecg_rate)

report(wilcox.test(df_ecg_noins$ecg_rate, df_ecg_gamif$ecg_rate, p.adj = "bonf"))
report(wilcox.test(df_ecg_gamif$ecg_rate, df_ecg_imita$ecg_rate, p.adj = "bonf"))
report(wilcox.test(df_ecg_noins$ecg_rate, df_ecg_imita$ecg_rate, p.adj = "bonf"))
