==> analysis/5.analysis_eda.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Based on discussions between [Changkun Ou](https://changkun.de), [Francesco Chiossi](https://www.um.informatik.uni-muenchen.de/personen/mitarbeiter/chiossi/index.html), and [Sven Mayer](https://sven-mayer.com).

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

source('stats.R')

df_eda <- read.csv("./clean/eda.csv")
# remove users with missing data, see 1.missing.ipynb
df_eda <- df_eda %>% filter(!(user_id %in% c(12, 14, 20, 22)))

df_eda$user_id <- as.factor(df_eda$user_id)
df_eda$method <- as.factor(df_eda$method)
df_eda$tide <- as.factor(df_eda$tide)

twoway_anova_test(df_eda, "tonic_avg")
twoway_anova_test(df_eda, "amplitude_avg")

df_eda_fall <- df_eda[df_eda$falls_num > 0,]
twoway_anova_test(df_eda_fall, "amplitude_prefall")
twoway_anova_test(df_eda_fall, "amplitude_postfall")
twoway_anova_test(df_eda_fall, "peaks_num")
twoway_anova_test(df_eda_fall, "peaks_prefall")
twoway_anova_test(df_eda_fall, "peaks_postfall")
twoway_anova_test(df_eda_fall, "falls_num")

df_eda_noins <- df_eda[df_eda$method == "NoInstructions",]
df_eda_gamif <- df_eda[df_eda$method == "Gamification",]
df_eda_imita <- df_eda[df_eda$method == "Imitation",]

mean(df_eda_noins$amplitude_avg)
sd(df_eda_noins$amplitude_avg)
mean(df_eda_gamif$amplitude_avg)
sd(df_eda_gamif$amplitude_avg)
mean(df_eda_imita$amplitude_avg)
sd(df_eda_imita$amplitude_avg)

report(wilcox.test(df_eda_noins$amplitude_avg, df_eda_gamif$amplitude_avg, p.adj = "bonf"))
report(wilcox.test(df_eda_gamif$amplitude_avg, df_eda_imita$amplitude_avg, p.adj = "bonf"))
report(wilcox.test(df_eda_noins$amplitude_avg, df_eda_imita$amplitude_avg, p.adj = "bonf"))
