==> analysis/6.analysis_motion.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Based on discussions between [Changkun Ou](https://changkun.de), [Francesco Chiossi](https://www.um.informatik.uni-muenchen.de/personen/mitarbeiter/chiossi/index.html), [Sven Mayer](https://sven-mayer.com), and Fabio Sarto.

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

source('stats.R')

df_motion <- read.csv("./clean/motion.csv")
# remove users with missing data, see 1.missing.ipynb
df_motion <- df_motion %>% filter(!(user_id %in% c(12, 14, 20, 22)))

df_motion$user_id <- as.factor(df_motion$user_id)
df_motion$method <- as.factor(df_motion$method)
df_motion$tide <- as.factor(df_motion$tide)

twoway_anova_test(df_motion, "duration")
twoway_anova_test(df_motion, "falls_num")
twoway_anova_test(df_motion, "mean_drift_head")
twoway_anova_test(df_motion, "std_drift_head")
twoway_anova_test(df_motion, "mean_drift_pelvis")
twoway_anova_test(df_motion, "std_drift_pelvis")
twoway_anova_test(df_motion, "mean_alpha")
twoway_anova_test(df_motion, "std_alpha")

df_motion_noins <- df_motion[df_motion$method == "NoInstructions",]
df_motion_gamif <- df_motion[df_motion$method == "Gamification",]
df_motion_imita <- df_motion[df_motion$method == "Imitation",]

mean(df_motion_noins$duration)
sd(df_motion_noins$duration)
mean(df_motion_gamif$duration)
sd(df_motion_gamif$duration)
mean(df_motion_imita$duration)
sd(df_motion_imita$duration)

report(wilcox.test(df_motion_noins$duration, df_motion_gamif$duration, p.adj = "bonf"))
report(wilcox.test(df_motion_gamif$duration, df_motion_imita$duration, p.adj = "bonf"))
report(wilcox.test(df_motion_noins$duration, df_motion_imita$duration, p.adj = "bonf"))
