==> analysis/7.analysis_ipq.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Based on discussions between [Changkun Ou](https://changkun.de), [Francesco Chiossi](https://www.um.informatik.uni-muenchen.de/personen/mitarbeiter/chiossi/index.html), and [Sven Mayer](https://sven-mayer.com).

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

source('stats.R')

df_paces <- read.csv("./clean/paces.csv")
df_paces <- df_paces %>% filter(!(user_id %in% c(12, 14, 20, 22)))
df_paces <- df_paces %>% group_by(user_id, method, tide) %>% summarise(tlx = mean(tlx))

df_paces$user_id <- as.factor(df_paces$user_id)
df_paces$method <- as.factor(df_paces$method)
df_paces$tide <- as.factor(df_paces$tide)

twoway_anova_test(df_paces, "tlx")

df_paces_noins <- df_paces[df_paces$method == "NoInstructions",]
df_paces_gamif <- df_paces[df_paces$method == "Gamification",]
df_paces_imita <- df_paces[df_paces$method == "Imitation",]

mean(df_paces_noins$tlx)
sd(df_paces_noins$tlx)
mean(df_paces_gamif$tlx)
sd(df_paces_gamif$tlx)
mean(df_paces_imita$tlx)
sd(df_paces_imita$tlx)

report(wilcox.test(df_paces_noins$tlx, df_paces_gamif$tlx, p.adj = "bonf"))
report(wilcox.test(df_paces_gamif$tlx, df_paces_imita$tlx, p.adj = "bonf"))
report(wilcox.test(df_paces_noins$tlx, df_paces_imita$tlx, p.adj = "bonf"))
