==> analysis/8.analysis_tlx.ipynb <==
Copyright © 2022 LMU Munich Media Informatics Group. All rights reserved.

Based on discussions between [Changkun Ou](https://changkun.de), [Francesco Chiossi](https://www.um.informatik.uni-muenchen.de/personen/mitarbeiter/chiossi/index.html), and [Sven Mayer](https://sven-mayer.com).

Use of this source code is governed by a GNU GPLv3 license that
can be found in the LICENSE file.

source('stats.R')

df_tlx <- read.csv("./clean/nasa_tlx.csv")
df_tlx <- df_tlx %>% filter(!(user_id %in% c(12, 14, 20, 22)))
df_tlx <- df_tlx %>% group_by(user_id, method, tide) %>% summarise(tlx = mean(tlx))

df_tlx$user_id <- as.factor(df_tlx$user_id)
df_tlx$method <- as.factor(df_tlx$method)
df_tlx$tide <- as.factor(df_tlx$tide)

twoway_anova_test(df_tlx, "tlx")

df_tlx_noins <- df_tlx[df_tlx$method == "NoInstructions",]
df_tlx_gamif <- df_tlx[df_tlx$method == "Gamification",]
df_tlx_imita <- df_tlx[df_tlx$method == "Imitation",]

mean(df_tlx_noins$tlx)
sd(df_tlx_noins$tlx)
mean(df_tlx_gamif$tlx)
sd(df_tlx_gamif$tlx)
mean(df_tlx_imita$tlx)
sd(df_tlx_imita$tlx)

report(wilcox.test(df_tlx_noins$tlx, df_tlx_gamif$tlx, p.adj = "bonf"))
report(wilcox.test(df_tlx_gamif$tlx, df_tlx_imita$tlx, p.adj = "bonf"))
report(wilcox.test(df_tlx_noins$tlx, df_tlx_imita$tlx, p.adj = "bonf"))
